GC2D/Guide linkSelect: u8 alpha local for pane alpha loop (96.16 -> 97.0)

+++ src/GC2D/Guide.cpp
@@ -674,7 +674,7 @@ void TGuide::linkSelect()
 		}
 	}
 
-	int alpha;
+	u8 alpha;
 	if ((int)unk160 < 0x1e) {
 		alpha = 0x1e;
 	} else if ((int)unk160 > 0xff) {
